backend/app/intel/scraper.py
"""Web scraper tool — hybrid httpx + Selenium approach.

OPTIMIZATION: Google News RSS XML is parsed with httpx/feedparser first
(no browser needed, ~80% less CPU). Selenium is only used as a fallback
for pages that require JavaScript rendering.

This prevents the Celery heartbeat timeouts caused by Selenium
consuming all available memory/CPU during dashboard requests.
"""

import asyncio
import uuid
import logging
from urllib.parse import quote
from datetime import datetime
from typing import Optional

import httpx
import feedparser
from dateutil import parser as dateparser

logger = logging.getLogger(__name__)

# Google News RSS base URL
GOOGLE_RSS_BASE = "https://news.google.com/rss/search?hl=en-IN&gl=IN&ceid=IN:en&q="

# Shared httpx client config
HTTPX_HEADERS = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
}
HTTPX_TIMEOUT = 10.0


def _parse_date(entry) -> Optional[datetime]:
    """Safely parse publish date from RSS entry."""
    try:
        if hasattr(entry, "published"):
            return dateparser.parse(entry.published)
    except Exception:
        pass
    return datetime.utcnow()


async def scrape_topic(query: str, max_articles: int = 8) -> list[dict]:
    """Scrape articles for a query using Google News RSS (no Selenium).

    This is the primary scraping path — lightweight, fast, and
    doesn't consume browser resources. Returns articles with
    title, summary, url, source, date.
    """
    url = GOOGLE_RSS_BASE + quote(query)
    articles = []

    try:
        async with httpx.AsyncClient(
            headers=HTTPX_HEADERS,
            timeout=HTTPX_TIMEOUT,
            follow_redirects=True,
        ) as client:
            resp = await client.get(url)
            resp.raise_for_status()

        feed = feedparser.parse(resp.text)

        for entry in feed.entries[:max_articles]:
            title = getattr(entry, "title", "")
            summary = getattr(entry, "summary", "")
            link = getattr(entry, "link", "")

            source_obj = getattr(entry, "source", None)
            if source_obj and hasattr(source_obj, "get"):
                source = source_obj.get("title", "Google News")
            elif hasattr(source_obj, "title"):
                source = source_obj.title
            else:
                source = "Google News"

            if title and link:
                articles.append({
                    "temp_id": str(uuid.uuid4()),
                    "title": title,
                    "summary": summary[:500] if summary else "",
                    "content": summary[:500] if summary else "",
                    "url": link,
                    "source": source,
                    "date": str(_parse_date(entry)),
                    "published": str(_parse_date(entry)),
                    "query_used": query,
                })

        logger.info(f"📡 RSS scraped '{query}': {len(articles)} articles (no browser)")

    except Exception as e:
        logger.warning(f"⚠️ RSS scrape failed for '{query}': {e}")
        # Fall back to Selenium only if RSS fails completely
        try:
            loop = asyncio.get_event_loop()
            articles = await loop.run_in_executor(
                None, _selenium_fallback, query, max_articles
            )
        except Exception as se:
            logger.error(f"❌ Selenium fallback also failed for '{query}': {se}")

    return articles


def _selenium_fallback(query: str, max_articles: int) -> list[dict]:
    """Last-resort Selenium scraper — only used when RSS parsing fails.

    This path is expensive (spawns Chrome) and should be rare.
    The RSS path handles 95%+ of queries successfully.
    """
    import os
    from bs4 import BeautifulSoup

    logger.info(f"🔧 Selenium fallback for: '{query}'")

    search_url = f"https://economictimes.indiatimes.com/searchresult.cms?query={quote(query)}"

    try:
        from selenium import webdriver
        from selenium.webdriver.chrome.options import Options
        from selenium.webdriver.chrome.service import Service

        options = Options()
        options.add_argument("--headless=new")
        options.add_argument("--disable-gpu")
        options.add_argument("--no-sandbox")
        options.add_argument("--disable-dev-shm-usage")
        options.add_argument("--disable-extensions")
        options.add_argument("--disable-images")
        # Memory limits to prevent OOM
        options.add_argument("--js-flags=--max-old-space-size=256")
        options.add_argument("--single-process")

        chrome_bin = os.environ.get("CHROME_BIN", "/usr/bin/chromium")
        chromedriver_path = os.environ.get("CHROMEDRIVER_PATH", "/usr/bin/chromedriver")
        options.binary_location = chrome_bin

        service = Service(chromedriver_path)
        driver = webdriver.Chrome(service=service, options=options)
        driver.set_page_load_timeout(12)  # tighter timeout
    except Exception as e:
        logger.error(f"Failed to init Chrome: {e}")
        return []

    try:
        driver.get(search_url)
    except Exception:
        try:
            driver.execute_script("window.stop();")
        except Exception:
            pass

    html = driver.page_source
    driver.quit()

    soup = BeautifulSoup(html, "html.parser")
    links = []

    for a in soup.find_all("a", href=True):
        href = a["href"]
        if "/articleshow/" in href:
            link = href if href.startswith("http") else f"https://economictimes.indiatimes.com{href}"
            if link not in links:
                links.append(link)

    articles = []
    for link in links[:max_articles]:
        article = _scrape_article_httpx(link)
        if article:
            article["query_used"] = query
            articles.append(article)

    return articles


def _scrape_article_httpx(url: str) -> dict | None:
    """Scrape a single article page using httpx (no browser)."""
    try:
        resp = httpx.get(url, headers=HTTPX_HEADERS, timeout=HTTPX_TIMEOUT, follow_redirects=True)
        from bs4 import BeautifulSoup
        soup = BeautifulSoup(resp.text, "html.parser")

        title_tag = soup.find("h1")
        if not title_tag:
            return None

        title = title_tag.text.strip()
        paragraphs = soup.find_all("p")
        content = " ".join([p.text.strip() for p in paragraphs])

        date = None
        time_tag = soup.find("time")
        if time_tag:
            date = time_tag.text.strip()

        if len(content) < 100:
            return None

        return {
            "temp_id": str(uuid.uuid4()),
            "title": title,
            "content": content[:800],
            "summary": content[:300],
            "date": date,
            "published": date or str(datetime.utcnow()),
            "source": "Economic Times",
            "url": url,
        }
    except Exception as e:
        logger.warning(f"Article scrape failed {url}: {e}")
        return None
